Move the “agents” folder under the “lib”
