+++ src/hooks/useSkills.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { BabyElfAGI } from '@/agents/babyelfagi/executer';
+import { BabyElfAGI } from '@/lib/agents/babyelfagi/executer';
 import { SPECIFIED_SKILLS } from '@/utils/constants';
 
 type SkillInfo = {
