+++ src/lib/agents/babyagi/service.ts
