+++ src/lib/agents/babyelfagi/registory/skillRegistry.ts
