+++ src/lib/agents/babyelfagi/skills/presets/directoryStructure.ts
