+++ src/lib/agents/babyelfagi/skills/presets/objectiveSaver.ts
