+++ src/lib/agents/babyelfagi/skills/presets/textCompletion.ts
