+++ src/lib/agents/babyelfagi/skills/skill.ts
