+++ src/lib/agents/babyelfagi/tools/utils/largeTextExtract.ts
