+++ src/lib/agents/babyelfagi/tools/utils/textCompletion.ts
