+++ src/lib/agents/base/AgentStream.ts
